localhost version of RaspberryPi setup

+++ admin/removeUser.php
@@ -2,31 +2,34 @@
     session_start();
     include("../dbInfo/database.php");
 
-    $userid = (int) $_GET["userid"];
+    if ($_SESSION["user"] == "admin") {
 
-    $connect = new mysqli($host, $user, $passwd, $db) or die("Spojení se nezdařilo");
-    $connect -> set_charset("UTF8") or die("Kódování nenastaveno");
-    
-    $SQL1 = $connect->prepare("DELETE FROM storagedata WHERE user_iduser = ?");
-    $SQL1->bind_param("i", $userid);
+        $userid = (int) htmlspecialchars($_GET["userid"]);
 
-    $SQL1->execute();
-    
-    
-    $SQL2 = $connect->prepare("DELETE FROM user WHERE iduser = ?");
-    $SQL2->bind_param("i",$userid);
+        $connect = new mysqli($host, $user, $passwd, $db) or die("Can't connect to db");
+        $connect -> set_charset("UTF8") or die("Encoding not set");
+        
+        $SQL1 = $connect->prepare("DELETE FROM storagedata WHERE user_iduser = ?");
+        $SQL1->bind_param("i", $userid);
 
-    $SQL2->execute();
+        $SQL1->execute();
+        
+        
+        $SQL2 = $connect->prepare("DELETE FROM user WHERE iduser = ?");
+        $SQL2->bind_param("i",$userid);
 
-    $connect->close();
+        $SQL2->execute();
 
-    $files = scandir($_SESSION["rootPath"] . $_GET["username"]);
-    var_dump($files);
+        $connect->close();
 
-    foreach($files as $file) {
-        unlink($_SESSION["rootPath"] . $_GET["username"] . "\\" . $file);
+        $files = scandir($_SESSION["rootPath"] . basename(htmlspecialchars($_GET["username"])));
+        var_dump($files);
+
+        foreach($files as $file) {
+            unlink($_SESSION["rootPath"] . basename(htmlspecialchars($_GET["username"])) . "/" . $file);
+        }
+
+        rmdir($_SESSION["rootPath"] . $_GET["username"]);
+
+        header("Location: index.php");
     }
-
-    rmdir($_SESSION["rootPath"] . $_GET["username"]);
-
-    header("Location: index.php");
